# Conflicts:
#	assets/translations/en.json
#	ios/Podfile.lock
#	lib/core/app.dart
#	lib/core/infrastructure/app_lifecycle_observer.dart
#	lib/core/infrastructure/dio_client/dio_client.dart
#	lib/core/presentation/alerting/modal_bottom_sheet.dart
#	lib/core/presentation/choice_chip/custom_choice_chip.dart
#	lib/core/presentation/error/error_screen.dart
#	lib/core/presentation/icon_images/app_icons.dart
#	lib/core/presentation/icon_images/app_images.dart
#	lib/core/presentation/localization/localized_texts.dart
#	lib/core/presentation/routes/app_router.dart
#	lib/features/account/presentation/account_page/account_page.dart
#	lib/features/account/presentation/account_page/widgets/account_section.dart
#	lib/features/account/presentation/account_page/widgets/emergency_btn.dart
#	lib/features/account/presentation/account_page/widgets/personal_details_section.dart
#	lib/features/account/presentation/account_page/widgets/preferences_section.dart
#	lib/features/account/presentation/account_page/widgets/report_abuse_section.dart
#	lib/features/account/presentation/account_page/widgets/section_item.dart
#	lib/features/account/presentation/account_page/widgets/section_title.dart
#	lib/features/account/presentation/emergency_numbers/emergency_number_card.dart
#	lib/features/authentication/application/dto/login_response.dart
#	lib/features/mood/presentation/create_mood_page.dart
#	lib/features/mood/presentation/mood_option_page.dart
#	lib/features/nutrition/presentation/log_weight_page/log_weight_page.dart
#	lib/features/physical_fitness/presentation/birthday/birthday_page.dart
#	lib/features/report_abuse/presentation/widget/report_abuse_widget.dart
#	lib/features/report_abuse/presentation/widget/report_section_title.dart
#	pubspec.lock
#	pubspec.yaml

+++ lib/features/nutrition/application/recipe/recipe_event.dart
@@ -15,7 +15,7 @@ class RecipeEvent with _$RecipeEvent {
 
   const factory RecipeEvent.servingChanged({
     required int mealId,
-    required int servingAmount,
+    required double servingAmount,
     required int recipeId,
   }) = ServingChanged;
